Address code feedback

+++ src/complex/DataStructure/NeighborList.cpp
@@ -90,7 +90,7 @@ int32 NeighborList<T>::eraseTuples(const std::vector<usize>& idxs)
   usize arraySize = m_Array.size();
   // Sanity Check the Indices in the vector to make sure we are not trying to remove any indices that are
   // off the end of the array and return an error code.
-  for(unsigned long idx : idxs)
+  for(usize idx : idxs)
   {
     if(idx >= arraySize)
     {
